+++ client/src/components/ItemCard/index.tsx
@@ -6,17 +6,6 @@ import Ratings from "../Ratings";
 
 import "./ItemCard.css";
 
-type ItemEntry = {
-  _id: string;
-  name: string;
-  description: string;
-  price: number;
-  avgRating: number;
-  isOnSale: boolean;
-  stockCount: number;
-  imageUrl: string;
-};
-
 type Props = {
   item: ItemEntry;
 };
