+++ client/src/pages/Deals.tsx
@@ -1,19 +1,7 @@
 import React, { useState, useEffect } from "react";
-import axios from "axios";
 
 import ItemContainer from "../components/ItemContainer";
 import API from "../utils/API";
-
-type ItemEntry = {
-  _id: string;
-  name: string;
-  description: string;
-  price: number;
-  avgRating: number;
-  isOnSale: boolean;
-  stockCount: number;
-  imageUrl: string;
-};
 
 const Deals: React.FC = () => {
   const [itemsList, setItemsList] = useState<ItemEntry[]>([]);
