+++ client/src/types/ItemEntry.d.ts
@@ -1,0 +1,10 @@
+type ItemEntry = {
+  _id: string;
+  name: string;
+  description: string;
+  price: number;
+  avgRating: number;
+  isOnSale: boolean;
+  stockCount: number;
+  imageUrl: string;
+};
